pyproject.toml
[tool.poetry]
name = "tutors-api"
version = "0.1.0"
description = ""
authors = ["uilic <uros.ilic@fundingoptions.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.0.1"
django-admin = "^2.0.1"
drf-yasg = "^1.20.0"
djangorestframework = "^3.13.1"
django-environ = "^0.8.1"
psycopg2 = "^2.9.3"
PyJWT = "1.7.1"
django-cors-headers = "^3.11.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
